Adds basic query-code syntax highlight changing
Adds the ability to modify the DOM to change the PrismJS class utilized
in the `query-code` block to update with the new class and remove the
previously used class prior to adding the new one. This can help to ease
formatting for visibilty as well as promote syntax definitions for the
other target query languages.